Make SSE action execution use an internal PubSub implementation
The mechanism introducted in this commit causes ingested ledgers to notify SSE action handlers,
letting them know they should query the database for new related information.
This causes Every SSE connection to query the history database only when a ledger was ingested
that contains explicit information related to that SSE request. This is in contrast to the
current way it's done - where every SSE action handler queries the database in a periodic manner
e.g. every 5s, regardless of that ledger's contents.

For example `/accounts/{account_id}/transactions` will query the history database only when
a ledger containing a transaction affecting that specific `{account_id}` is ingested.

This greatly reduces the performance load on the history database belonging to Horizons
handling a large number of SSE connections.

In addition:

- Add a new `GetPubsubTopic()` function to `actions.EventStreamer` interface.
  It returns the specific URL parameter used to notify on related ledger ingestions.
- Remove `sse-update-frequency` command line argument, since the feature introduced in this
  commit makes it obsolete.
- Added new `services/horizon/internal/sse_test.go` with related unit tests.